+++ packages/apollo-server-express/src/ApolloServer.ts
@@ -4,8 +4,6 @@ import { json, OptionsJson } from 'body-parser';
 import {
   GraphQLOptions,
   ApolloServerBase,
-  ContextFunction,
-  Context,
   Config,
   runHttpQuery,
   HttpQueryError,
@@ -41,15 +39,11 @@ export interface ExpressContext {
   res: express.Response;
 }
 
-export interface ApolloServerExpressConfig extends Config {
-  context?: ContextFunction<ExpressContext, Context> | Context;
-}
+export type ApolloServerExpressConfig = Config<ExpressContext>;
 
-export class ApolloServer extends ApolloServerBase {
-  constructor(config: ApolloServerExpressConfig) {
-    super(config);
-  }
-
+export class ApolloServer<
+  ContextFunctionParams = ExpressContext,
+> extends ApolloServerBase<ContextFunctionParams> {
   // This translates the arguments from the middleware into graphQL options It
   // provides typings for the integration specific behavior, ideally this would
   // be propagated with a generic to the super class
@@ -57,7 +51,8 @@ export class ApolloServer extends ApolloServerBase {
     req: express.Request,
     res: express.Response,
   ): Promise<GraphQLOptions> {
-    return super.graphQLServerOptions({ req, res });
+    const contextParams: ExpressContext = { req, res };
+    return super.graphQLServerOptions(contextParams);
   }
 
   public applyMiddleware({ app, ...rest }: ServerRegistration) {